Small correction + comment added.

+++ invade/CalculatorFinance.hs
@@ -218,12 +218,15 @@ calcCommission  account market stockname price shares = do
     var_market <- peekCString market
     var_stockname <- peekCString stockname
     case lowerCase var_account of
-        "binb00" -> return (realToFrac $ getBinb00Commission var_market var_stockname $ realToFrac amount_simple)
+        "binb00" -> return (realToFrac $ getBinb00Commission var_market var_stockname $ amount_simple)
         "whsi00" -> return (realToFrac $ getWhsi00Commission var_market var_stockname (realToFrac price) (fromIntegral shares))
         _ -> return (fromIntegral 0.0)
     where
+        -- TODO: calcAmountSimple: we try to take a CDouble out
+        -- of an IO CDouble, to realToFrac it to a Double.
+        -- Test if it works this way and/or find the correct way to do it.
         amount_simple = do
-            realToFrac . foldl . calcAmountSimple (realToFrac price) (fromIntegral shares)
+            realToFrac . foldl . calcAmountSimple price shares
 
 -- TODO: get 2500 etc values from T_PARAMETER
 getBinb00Commission :: String -> String -> Double -> Double
